Added Multiply and Modulo functions
Added implementation of Multiply and Modulo functions.

+++ src/math_lib/mathematical_library.h
@@ -47,4 +47,4 @@ int Factorial(int operand);
  * @param rightOperand - divisor
  * @return result
  */
-double Modulo(int leftOperand, int rightOperand);
+int Modulo(int leftOperand, int rightOperand);
